+++ lib/localization/manuel_route_builder_default_l10n.dart
@@ -38,7 +38,8 @@ class ManuelRouteBuilderDefaultL10n implements ManuelRouteBuilderL10n {
   String get selectAreaOnMap => "Select area on map";
 
   @override
-  String get drawCircleByClickingOnMap => "Draw a circle by clicking on the map.";
+  String get drawCircleByClickingOnMap =>
+      "Draw a circle by clicking on the map.";
 
   @override
   String get freeDrawOnMap => "Free draw on map";
@@ -83,6 +84,7 @@ class ManuelRouteBuilderDefaultL10n implements ManuelRouteBuilderL10n {
   String get buildRoute => "Build route";
 
   @override
-  String pointsCountInArea(int count) => "$count point${count > 1 ? 's' : ''} in this area";
+  String pointsCountInArea(int count) =>
+      "$count point${count > 1 ? 's' : ''} in this area";
   String pointsCount(int count) => "$count point${count > 1 ? 's' : ''}";
 }